flatten invoice
