+++ src/ui/invoice/design.tsx
@@ -1,1 +1,29 @@
-export * from './designs/basic/invoice'
+'use client'
+
+import {
+  Attachment as AttachmentPages,
+  Invoice as RecordPages,
+} from '~/ui/invoice/designs/basic/invoice'
+
+import { AttachmentProvider } from '~/ui/hooks/use-attachment'
+import { RecordProvider, useRecord } from '~/ui/hooks/use-record'
+
+export function Invoice() {
+  let record = useRecord()
+
+  return (
+    <RecordProvider record={record}>
+      <div className="grid w-full grid-flow-row gap-8 font-pdf print:gap-0">
+        <RecordPages />
+
+        {record.attachments.map((attachment) => {
+          return (
+            <AttachmentProvider key={attachment.id} document={attachment}>
+              <AttachmentPages />
+            </AttachmentProvider>
+          )
+        })}
+      </div>
+    </RecordProvider>
+  )
+}
